requirements.txt
dataclasses-json==0.6.7
discord-py-interactions==5.12.1
requests==2.32.3
python-dotenv>=0.19.2
py-dactyl>=2.0.0
aiohttp>=3.9.0